src/MOTChallengeEvalKit/MOT/requirements.txt
matlabengineforpython===R2020a
numpy==1.19.0
opencv-python==4.2.0.34
pandas==1.0.5
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
